+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 0.3.0
+
+- Add a setting to reroll group initiatives at the start of each combat round
+- Prevent the 'G' keyboard shortcut from triggering an initiative reroll once combat has started
+- If the active encounter already has combatants, you no longer need to select tokens on the current scene to trigger a group initiative roll
+
 ## 0.2.0
 
 - Add a setting to configure whether player tokens should be grouped with friendly tokens